Split arbitraries
