+++ modules/seqexec/server/src/test/scala/seqexec/server/ExecutionQueueSpec.scala
@@ -5,10 +5,9 @@ package seqexec.server
 
 import cats.tests.CatsSuite
 import monocle.law.discipline.LensTests
-import SeqexecServerArbitraries._
 import seqexec.model.SeqexecModelArbitraries._
 
-final class ExecutionQueueSpec extends CatsSuite {
+final class ExecutionQueueSpec extends CatsSuite with SeqexecServerArbitraries {
   checkAll("ExecutionQueue name lens", LensTests(ExecutionQueue.name))
   checkAll("ExecutionQueue command state lens",
            LensTests(ExecutionQueue.cmdState))
