+++ modules/seqexec/server/src/test/scala/seqexec/server/gnirs/GnirsSpec.scala
@@ -11,9 +11,7 @@ import edu.gemini.spModel.gemini.gnirs.GNIRSParams._
 /**
   * Tests GNIRS typeclasses
   */
-final class GnirsSpec extends CatsSuite {
-  import seqexec.server.SeqexecServerArbitraries._
-
+final class GnirsSpec extends CatsSuite with GnirsArbitraries {
   checkAll("Eq[AcquisitionMirror]", EqTests[AcquisitionMirror].eqv)
   checkAll("Eq[WollastonPrism]", EqTests[WollastonPrism].eqv)
   checkAll("Eq[SlitWidth]", EqTests[SlitWidth].eqv)
