+++ modules/seqexec/server/src/test/scala/seqexec/server/keywords/KeywordsSpec.scala
@@ -5,9 +5,10 @@ package seqexec.server.keywords
 
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline._
-import seqexec.server.SeqexecServerArbitraries._
+import gem.enum.KeywordName
 
-final class KeywordsSpec extends CatsSuite {
+final class KeywordsSpec extends CatsSuite with KeywordArbitraries {
+  checkAll("Eq[KeywordName]", EqTests[KeywordName].eqv)
   checkAll("Eq[KeywordType]", EqTests[KeywordType].eqv)
   checkAll("Eq[InternalKeyword]", EqTests[InternalKeyword].eqv)
   checkAll("Eq[KeywordBag]", EqTests[KeywordBag].eqv)
